+++ crypto/src/lib.rs
@@ -200,13 +200,13 @@ pub trait CryptoHash {
 /// ```
 pub fn init() {
     if !crypto_impl::init() {
-        panic!("Cryptographic library hasn't initialized.");
+        panic!("Cryptographic library initialization failed.");
     }
 }
 
 /// This structure provides a possibility to calculate a hash digest
 /// for a stream of data. Unlike the
-/// [`Hash` structure](https://docs.rs/exonum/0.7.0/exonum/crypto/struct.Hash.html),
+/// [`Hash` structure](struct.Hash.html),
 /// the given structure lets the code process several data chunks without
 /// the need to copy them into a single buffer.
 ///