+++ exonum/src/api/mod.rs
@@ -24,6 +24,8 @@ use std::{collections::BTreeMap, fmt};
 
 use self::{backends::actix, node::public::ExplorerApi};
 use blockchain::{Blockchain, SharedNodeState};
+use crypto::PublicKey;
+use node::ApiSender;
 
 pub mod backends;
 pub mod error;
@@ -218,6 +220,7 @@ impl ServiceApiScope {
 /// ```
 #[derive(Debug, Clone, Default)]
 pub struct ServiceApiBuilder {
+    blockchain: Option<Blockchain>,
     public_scope: ServiceApiScope,
     private_scope: ServiceApiScope,
 }
@@ -225,7 +228,17 @@ pub struct ServiceApiBuilder {
 impl ServiceApiBuilder {
     /// Creates a new service API builder.
     pub fn new() -> Self {
-        Self::default()
+        Self {
+            blockchain: None,
+            ..Default::default()
+        }
+    }
+
+    fn with_blockchain(blockchain: Blockchain) -> Self {
+        Self {
+            blockchain: Some(blockchain),
+            ..Default::default()
+        }
     }
 
     /// Returns a mutable reference to the public API scope builder.
@@ -236,6 +249,22 @@ impl ServiceApiBuilder {
     /// Returns a mutable reference to the private API scope builder.
     pub fn private_scope(&mut self) -> &mut ServiceApiScope {
         &mut self.private_scope
+    }
+
+    /// Returns an optional reference to the Blockchain.
+    pub fn blockchain(&self) -> Option<&Blockchain> {
+        self.blockchain.as_ref()
+    }
+
+    /// Returns an optional reference to the ApiSender.
+    pub fn api_sender(&self) -> Option<&ApiSender> {
+        self.blockchain().map(|blockchain| &blockchain.api_sender)
+    }
+
+    /// Returns an optional value to the PublicKey.
+    pub fn public_key(&self) -> Option<PublicKey> {
+        self.blockchain()
+            .map(|blockchain| blockchain.service_keypair.0)
     }
 }
 
@@ -292,7 +321,7 @@ impl ApiAggregator {
         );
         // Adds services APIs.
         inner.extend(blockchain.service_map().iter().map(|(_, service)| {
-            let mut builder = ServiceApiBuilder::new();
+            let mut builder = ServiceApiBuilder::with_blockchain(blockchain.clone());
             service.wire_api(&mut builder);
             // TODO think about prefixes for non web backends. (ECR-1758)
             let prefix = format!("services/{}", service.service_name());