+++ exonum/src/events/noise/tests.rs
@@ -37,7 +37,6 @@ use events::{
     tests::raw_message,
 };
 use messages::RawMessage;
-use node::state::SharedConnectList;
 
 #[test]
 #[cfg(feature = "sodiumoxide-crypto")]
@@ -176,21 +175,17 @@ fn check_encrypt_decrypt_message(msg_size: usize) {
 }
 
 fn create_noise_sessions() -> (NoiseWrapper, NoiseWrapper) {
-    let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new([1; SEED_LENGTH]));
-
-    let mut params =
-        HandshakeParams::new(public_key, secret_key, SharedConnectList::default(), 1024);
-    params.set_remote_key(public_key);
+    let params = HandshakeParams::with_default_params();
 
     let mut initiator = NoiseWrapper::initiator(&params);
     let mut responder = NoiseWrapper::responder(&params);
 
-    let buffer_out = initiator.write_handshake_msg().unwrap();
+    let buffer_out = initiator.write_handshake_msg(&[]).unwrap();
     responder.read_handshake_msg(&buffer_out).unwrap();
 
-    let buffer_out = responder.write_handshake_msg().unwrap();
+    let buffer_out = responder.write_handshake_msg(&[]).unwrap();
     initiator.read_handshake_msg(&buffer_out).unwrap();
-    let buffer_out = initiator.write_handshake_msg().unwrap();
+    let buffer_out = initiator.write_handshake_msg(&[]).unwrap();
     responder.read_handshake_msg(&buffer_out).unwrap();
 
     (
@@ -239,19 +234,11 @@ const MAX_MESSAGE_LEN: usize = 128;
 const EMPTY_MESSAGE: &[u8] = &[0; 0];
 const STANDARD_MESSAGE: &[u8] = &[0; MAX_MESSAGE_LEN];
 
-pub fn default_test_params() -> HandshakeParams {
-    let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new([1; SEED_LENGTH]));
-    let mut params =
-        HandshakeParams::new(public_key, secret_key, SharedConnectList::default(), 1024);
-    params.set_remote_key(public_key);
-    params
-}
-
 #[test]
 #[should_panic(expected = "WrongMessageLength")]
 fn test_noise_handshake_errors_ee_empty() {
     let addr: SocketAddr = "127.0.0.1:45003".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::EphemeralKeyExchange,
         EMPTY_MESSAGE,
@@ -265,7 +252,7 @@ fn test_noise_handshake_errors_ee_empty() {
 #[should_panic(expected = "WrongMessageLength")]
 fn test_noise_handshake_errors_es_empty() {
     let addr: SocketAddr = "127.0.0.1:45004".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::StaticKeyExchange,
         EMPTY_MESSAGE,
@@ -279,7 +266,7 @@ fn test_noise_handshake_errors_es_empty() {
 #[should_panic(expected = "Dh")]
 fn test_noise_handshake_errors_ee_standard() {
     let addr: SocketAddr = "127.0.0.1:45005".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::EphemeralKeyExchange,
         STANDARD_MESSAGE,
@@ -293,7 +280,7 @@ fn test_noise_handshake_errors_ee_standard() {
 #[should_panic(expected = "Decrypt")]
 fn test_noise_handshake_errors_es_standard() {
     let addr: SocketAddr = "127.0.0.1:45006".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::StaticKeyExchange,
         STANDARD_MESSAGE,
@@ -307,7 +294,7 @@ fn test_noise_handshake_errors_es_standard() {
 #[should_panic(expected = "WrongMessageLength")]
 fn test_noise_handshake_errors_ee_empty_listen() {
     let addr: SocketAddr = "127.0.0.1:45007".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::EphemeralKeyExchange,
         EMPTY_MESSAGE,
@@ -321,7 +308,7 @@ fn test_noise_handshake_errors_ee_empty_listen() {
 #[should_panic(expected = "Dh")]
 fn test_noise_handshake_errors_ee_standard_listen() {
     let addr: SocketAddr = "127.0.0.1:45008".parse().unwrap();
-    let params = default_test_params();
+    let params = HandshakeParams::with_default_params();
     let bogus_message = Some(BogusMessage::new(
         HandshakeStep::EphemeralKeyExchange,
         STANDARD_MESSAGE,
@@ -335,7 +322,7 @@ fn test_noise_handshake_errors_ee_standard_listen() {
 #[should_panic(expected = "Decrypt")]
 fn test_noise_handshake_wrong_remote_key() {
     let addr: SocketAddr = "127.0.0.1:45009".parse().unwrap();
-    let mut params = default_test_params();
+    let mut params = HandshakeParams::with_default_params();
     let (remote_key, _) = gen_keypair_from_seed(&Seed::new([2; SEED_LENGTH]));
     params.set_remote_key(remote_key);
 
@@ -464,7 +451,7 @@ impl NoiseErrorHandshake {
 
         inner
             .read_handshake_msg(stream)
-            .map(move |(stream, inner)| {
+            .map(move |(stream, inner, _)| {
                 self.inner = Some(inner);
                 (stream, self)
             })
@@ -492,7 +479,7 @@ impl NoiseErrorHandshake {
 
             Either::B(
                 inner
-                    .write_handshake_msg(stream)
+                    .write_handshake_msg(stream, &[])
                     .map(move |(stream, inner)| {
                         self.inner = Some(inner);
                         self.current_step = self.current_step
@@ -513,7 +500,7 @@ impl Handshake for NoiseErrorHandshake {
         let framed = self.read_handshake_msg(stream)
             .and_then(|(stream, handshake)| handshake.write_handshake_msg(stream))
             .and_then(|(stream, handshake)| handshake.read_handshake_msg(stream))
-            .and_then(|(stream, handshake)| handshake.inner.unwrap().finalize(stream));
+            .and_then(|(stream, handshake)| handshake.inner.unwrap().finalize(stream, Vec::new()));
         Box::new(framed)
     }
 
@@ -524,7 +511,7 @@ impl Handshake for NoiseErrorHandshake {
         let framed = self.write_handshake_msg(stream)
             .and_then(|(stream, handshake)| handshake.read_handshake_msg(stream))
             .and_then(|(stream, handshake)| handshake.write_handshake_msg(stream))
-            .and_then(|(stream, handshake)| handshake.inner.unwrap().finalize(stream));
+            .and_then(|(stream, handshake)| handshake.inner.unwrap().finalize(stream, Vec::new()));
         Box::new(framed)
     }
 }